+++ R/plotRarefaction.r
@@ -32,7 +32,7 @@
 #' fractional time slices ranging from 1% to 100%.
 #' For each slice width *w = fT*, where *f* is a fraction and *T* is the total
 #' observation time for the individuals, random time windows of length *w* are repeatedly drawn
-#' across the individual’s observation period.
+#' across the individual's observation period.
 #' A behaviour counts as observed in that window if its duration overlaps with the window in time.
 #'
 #' For each time slice, the function computes the mean number of unique behaviours and the
@@ -54,13 +54,16 @@
 #' plotRarefaction(dat)
 #' }
 #'
+#' @importFrom grDevices adjustcolor hcl.colors
+#' @importFrom graphics polygon segments mtext lines legend
 #' @export
 plotRarefaction <- function(
-    x,
-    zoo = c("Zoo", "Beekse", "Safaripark", "Tiergarten"),
-    age = c("Mládě", "Zyqarri"),
-    cols = c("#3C8ABF", "#A1BCD7", "#768D1A", "#B1BE94"),
-    nBoot = 500) {
+  x,
+  zoo = c("Zoo", "Beekse", "Safaripark", "Tiergarten"),
+  age = c("Mládě", "Zyqarri"),
+  cols = c("#3C8ABF", "#A1BCD7", "#768D1A", "#B1BE94"),
+  nBoot = 500
+) {
   # remove non-behaviour rows
   x <- x[x$Behavior != "Mimo dohled", ]